modified:   bt vong lap/8d.cpp

+++ bt vong lap/8d.cpp
@@ -7,13 +7,13 @@ using namespace std;
 
 int main()
 {
-    int n, sum = 0;
+    uint64_t n, sum = 0;
     cin >> n;
-    int oldDigit = n % 10;
+    uint64_t oldDigit = n % 10;
     n /= 10;
     while(n != 0)
     {
-        int digit = n % 10;
+        uint64_t digit = n % 10;
         sum += abs(digit - oldDigit);
         oldDigit = digit;
         n /= 10;
